Synchronize with NotepadPlusPlusPluginPack.Net
Synchronize with NotepadPlusPlusPluginPack.Net to eliminate the possibility that code difference was causing the crashing issue #67 (it was not)

+++ CSVLintNppPlugin/PluginInfrastructure/Scintilla_iface.cs
@@ -3230,7 +3230,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
         /// <param name="cpmin">range to search</param>
         /// <param name="cpmax">range to search</param>
         /// <param name="searchText">the search pattern</param>
-        public TextToFind(int cpmin, int cpmax, string searchText)
+        public TextToFind(IntPtr cpmin, IntPtr cpmax, string searchText)
         {
             _sciTextToFind.chrg.cpMin = cpmin;
             _sciTextToFind.chrg.cpMax = cpmax;
